backend/requirements.txt
python-dotenv
aiohttp
fastapi[all]
uvicorn
semantic_kernel